% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_cache}
\alias{build_cache}
\title{Build the cache for \code{depsversion}}
\usage{
build_cache(
  pkg,
  ...,
  cache = get_cache_file(),
  target = getOption("depsversion_target", "./miniCRAN"),
  verbose = TRUE,
  update_cache = TRUE,
  dependencies = TRUE
)
}
\arguments{
\item{pkg}{the packages to}

\item{cache}{path to a \code{.rds} file used as a cache.}

\item{target}{path to a directory used as a local repository. Can take several Gb of space.}

\item{verbose}{whether to print progress}

\item{update_cache}{whether to update the \code{cache} file. Used for debugging.}

\item{dependencies}{whether to import dependencies of \code{pkg}}
}
\value{
invisibly, the tibble saved to \code{cache}
}
\description{
Build the cache for \code{depsversion}
}
