% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depsversion-package.R
\docType{package}
\name{depsversion-package}
\alias{depsversion}
\alias{depsversion-package}
\title{depsversion: Finds the minimum version of dependencies}
\description{
Finds the minimum version of dependencies needed in a package by checking when imported function were exported.
}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gmail.com} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

}
\keyword{internal}
