% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get_code}
\alias{get_code}
\title{Title}
\usage{
get_code(
  ns_path,
  cache = getOption("depsversion_cache", "data_ns.rds"),
  verbose = TRUE
)
}
\arguments{
\item{ns_path}{path to NAMESPACE}

\item{verbose}{whether to print the result}

\item{data}{the result of \link{build_cache} or its cache file}
}
\value{
the code to include in DESCRIPTION (invisibly if \code{verbose==TRUE})
}
\description{
Title
}
