% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_github.R
\name{update_cache_github}
\alias{update_cache_github}
\title{Update cache from GitHub}
\usage{
update_cache_github(cache = get_cache_file(), backup = TRUE, verbose = TRUE)
}
\arguments{
\item{cache}{the local cache file}

\item{backup}{whether to backup the local cache file}

\item{verbose}{whether to print stuff}
}
\value{
invisibly the new cache data
}
\description{
Update the cache used by \code{depversion} using a file hosted on GitHub.
}
